requirements.txt
faiss-cpu==1.11.0
Flask==3.1.1
langchain==0.3.26
langchain_community==0.3.26
langchain_groq==0.3.2
pymongo==4.13.2
PyPDF2==3.0.1
python-dotenv==1.1.0
redis==6.2.0
Werkzeug==3.1.3
